+++ tests/testthat/test_printf.R
@@ -15,7 +15,6 @@ test_that("catf into file", {
 })
 
 
-
 test_that("warningf", {
   expect_warning(warningf("xxx%ixxx", 123), "xxx123xxx")
   f = function() warningf("123")
@@ -31,22 +30,8 @@ test_that("stopf", {
   expect_error(stopf("abc"), "abc")
 })
 
-test_that("non-leanified call is printed", {
-  A = R6Class("A", public = list(warn = function() warningf("test warning")))
-  leanify_r6(A)
-  a = A$new()
-
-  tryCatch(a$warn(),
-    warning = function(w) {
-      expect_equal(w$call, quote(a$warn()))
-    }
-  )
-
-  f = function() warningf("test warning")
-
-  tryCatch(f(),
-    warning = function(w) {
-      expect_equal(w$call, quote(f()))
-    }
-  )
+test_that("formatting", {
+  expect_snapshot(stopf("abc"), error = TRUE)
+  expect_snapshot(stopf("s: %s", "b"), error = TRUE)
+  expect_snapshot(warningf("abc"))
 })